Reproduce HTML markup using <AddReaction> in React
