+++ templates/repo/issue/view_content/add_reaction.tmpl
@@ -1,16 +1,11 @@
 {{if .ctx.IsSigned}}
-<div class="item action ui pointing select-reaction dropdown top right" data-action-url="{{ .ActionURL }}">
-	<a class="add-reaction">
-		{{svg "octicon-smiley"}}
-	</a>
-	<div class="menu">
-		<div class="header">{{ .ctx.i18n.Tr "repo.pick_reaction"}}</div>
-		<div class="divider"></div>
-
-    <div class="react" data-react="add_reaction"></div>
-		{{range $value := AllowedReactions}}
-			<div class="item reaction" data-content="{{$value}}">{{ReactionToEmoji $value}}</div>
-		{{end}}
-	</div>
-</div>
+  <div
+    class="react"
+    data-react="add_reaction"
+    data-action-url="{{ .ActionURL }}"
+    data-i18n-pick="{{ .ctx.i18n.Tr "repo.pick_reaction"}}"
+    data-octicon="{{svg "octicon-smiley"}}"
+  >
+    {{svg "octicon-smiley"}}
+  </div>
 {{end}}
